requirements.txt
# Core FastAPI framework and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Environment variables
python-dotenv==1.0.0

# MQTT client for IoT communication
paho-mqtt==1.6.1

# Data validation and serialization
pydantic==2.5.0

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing framework and client
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Optional: Production server (alternative to uvicorn)
gunicorn==21.2.0

# Optional: Database drivers (uncomment as needed)
# psycopg2-binary==2.9.9  # PostgreSQL
# pymysql==1.1.0          # MySQL
# sqlite3                 # Built into Python (SQLite)

# Development dependencies (optional)
# black==23.11.0          # Code formatter
# flake8==6.1.0           # Linter
# mypy==1.7.1             # Type checker